refactor: confrom src directory
