+++ packages/location-state-conform/src/index.ts
@@ -1,1 +1,1 @@
-export * from "./use-location-form";
+export * from "./hooks";
